Welcome.us
In January 2022,  I completed a design audit for a newly formed nonprofit. on behalf of US Digital Response. I audited this pro bono client's website to give design feedback ahead of a planned advertising campaign during the Super Bowl and Olympics this past winter, with the goal to improve the user experience so new website visitors could get where they need to go on the website in less time. I delivered personas, journey maps, and a prioritized to-do list as part of this audit. 

My key recommendation -- to replace the email capture button on the homepage with several specific buttons telling users which actions they should take -- was implemented on the website within weeks of the audit.